+++ ShellKrypt.Core/CryptoTools/IPasswordGenerator.cs
@@ -1,0 +1,6 @@
+namespace ShellKrypt.Core.CryptoTools;
+
+public interface IPasswordGenerator
+{
+    string? GeneratePassword(PasswordGenerationOptions options);
+}