+++ ShellKrypt.Desktop/Views/AfterUnlock/CryptoTools/CryptoToolsView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace ShellKrypt.Desktop.Views.AfterUnlock.CryptoTools;
+
+public partial class CryptoToolsView : UserControl
+{
+    public CryptoToolsView()
+    {
+        InitializeComponent();
+    }
+}